Added tests for TokenRequester refactor
