+++ __tests__/login/oidc/redirectHandler/AuthCodeRedirectHandler.spec.ts
@@ -34,26 +34,24 @@ import {
 import AuthCodeRedirectHandler from "../../../../src/login/oidc/redirectHandler/AuthCodeRedirectHandler";
 import { RedirectorMock } from "../../../../src/login/oidc/__mocks__/Redirector";
 import { Response as NodeResponse } from "node-fetch";
+import { SessionCreatorMock } from "../../../../src/solidSession/__mocks__/SessionCreator";
+import { TokenRequesterMock } from "../../../../src/login/oidc/__mocks__/TokenRequester";
 
 describe("AuthCodeRedirectHandler", () => {
   const defaultMocks = {
     storageUtility: StorageUtilityMock,
-    issuerConfigFetcher: IssuerConfigFetcherMock,
-    fetcher: FetcherMock,
-    dpopHeaderCreator: DpopHeaderCreatorMock,
-    tokenSaver: TokenSaverMock,
-    redirector: RedirectorMock
+    redirector: RedirectorMock,
+    tokenRequester: TokenRequesterMock,
+    sessionCreator: SessionCreatorMock
   };
   function getAuthCodeRedirectHandler(
     mocks: Partial<typeof defaultMocks> = defaultMocks
   ): AuthCodeRedirectHandler {
     return new AuthCodeRedirectHandler(
       mocks.storageUtility ?? defaultMocks.storageUtility,
-      mocks.issuerConfigFetcher ?? defaultMocks.issuerConfigFetcher,
-      mocks.fetcher ?? defaultMocks.fetcher,
-      mocks.dpopHeaderCreator ?? defaultMocks.dpopHeaderCreator,
-      mocks.tokenSaver ?? defaultMocks.tokenSaver,
-      mocks.redirector ?? defaultMocks.redirector
+      mocks.redirector ?? defaultMocks.redirector,
+      mocks.tokenRequester ?? defaultMocks.tokenRequester,
+      mocks.sessionCreator ?? defaultMocks.sessionCreator
     );
   }
 
@@ -95,128 +93,41 @@ describe("AuthCodeRedirectHandler", () => {
     it("Makes a code request to the correct place", async () => {
       defaultMocks.storageUtility.getForUser
         .mockResolvedValueOnce("a")
-        .mockResolvedValueOnce("b")
-        .mockResolvedValueOnce("c")
-        .mockResolvedValueOnce("d")
-        .mockResolvedValueOnce(null);
-      defaultMocks.fetcher.fetch.mockResolvedValueOnce(
-        /* eslint-disable @typescript-eslint/camelcase */
-        (new NodeResponse(
-          JSON.stringify({
-            id_token: "idToken",
-            access_token: "accessToken"
-          })
-        ) as unknown) as Response
-        /* eslint-enable @typescript-eslint/camelcase */
-      );
+        .mockResolvedValueOnce("b");
       const authCodeRedirectHandler = getAuthCodeRedirectHandler();
       await authCodeRedirectHandler.handle(
         "https://coolsite.com/?code=someCode&state=userId"
       );
-      expect(defaultMocks.fetcher.fetch).toBeCalledWith(
-        IssuerConfigFetcherFetchConfigResponse.tokenEndpoint,
-        {
-          method: "POST",
-          headers: {
-            DPoP: DpopHeaderCreatorResponse,
-            "content-type": "application/x-www-form-urlencoded"
-          },
-          body:
-            "client_id=c&grant_type=authorization_code&code_verifier=a&code=someCode&redirect_uri=d"
-        }
-      );
-      expect(
-        defaultMocks.tokenSaver.saveTokenAndGetSession
-      ).toHaveBeenCalledWith("userId", "idToken", "accessToken", undefined);
-    });
-
-    it("throws an error if the idp returns bad data", async () => {
-      defaultMocks.storageUtility.getForUser
-        .mockResolvedValueOnce("a")
-        .mockResolvedValueOnce("b")
-        .mockResolvedValueOnce("c")
-        .mockResolvedValueOnce("d")
-        .mockResolvedValueOnce(null);
-      defaultMocks.fetcher.fetch.mockResolvedValueOnce(
+      expect(defaultMocks.tokenRequester.request).toHaveBeenCalledWith(
+        "userId",
         /* eslint-disable @typescript-eslint/camelcase */
-        (new NodeResponse(
-          JSON.stringify({
-            id_token: "onTheIdTokenIsHere"
-          })
-        ) as unknown) as Response
+        {
+          code: "someCode",
+          code_verifier: "a",
+          grant_type: "authorization_code",
+          redirect_uri: "b"
+        }
         /* eslint-enable @typescript-eslint/camelcase */
       );
+      expect(defaultMocks.redirector.redirect).toHaveBeenCalledWith(
+        "https://coolsite.com/",
+        {
+          redirectByReplacingState: true
+        }
+      );
+    });
+
+    it("Fails if a session was not retrieved", async () => {
+      defaultMocks.storageUtility.getForUser
+        .mockResolvedValueOnce("a")
+        .mockResolvedValueOnce("b");
+      defaultMocks.sessionCreator.getSession.mockResolvedValueOnce(null);
       const authCodeRedirectHandler = getAuthCodeRedirectHandler();
       await expect(
         authCodeRedirectHandler.handle(
           "https://coolsite.com/?code=someCode&state=userId"
         )
-      ).rejects.toThrowError("IDP token route returned an invalid response.");
-    });
-
-    it("throws an error if the idp provides a refresh token that is not a string", async () => {
-      defaultMocks.storageUtility.getForUser
-        .mockResolvedValueOnce("a")
-        .mockResolvedValueOnce("b")
-        .mockResolvedValueOnce("c")
-        .mockResolvedValueOnce("d")
-        .mockResolvedValueOnce(null);
-      defaultMocks.fetcher.fetch.mockResolvedValueOnce(
-        /* eslint-disable @typescript-eslint/camelcase */
-        (new NodeResponse(
-          JSON.stringify({
-            id_token: "idToken",
-            access_token: "access_token",
-            refresh_token: { this: "is not a string" }
-          })
-        ) as unknown) as Response
-        /* eslint-enable @typescript-eslint/camelcase */
-      );
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler();
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://coolsite.com/?code=someCode&state=userId"
-        )
-      ).rejects.toThrowError("IDP token route returned an invalid response.");
-    });
-
-    it("makes a code request with a Basic authorization header if a secret is present", async () => {
-      defaultMocks.storageUtility.getForUser
-        .mockResolvedValueOnce("a")
-        .mockResolvedValueOnce("b")
-        .mockResolvedValueOnce("c")
-        .mockResolvedValueOnce("d")
-        .mockResolvedValueOnce("e");
-      defaultMocks.fetcher.fetch.mockResolvedValueOnce(
-        /* eslint-disable @typescript-eslint/camelcase */
-        (new NodeResponse(
-          JSON.stringify({
-            id_token: "idToken",
-            access_token: "accessToken"
-          })
-        ) as unknown) as Response
-        /* eslint-enable @typescript-eslint/camelcase */
-      );
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler();
-      await authCodeRedirectHandler.handle(
-        "https://coolsite.com/?code=someCode&state=userId"
-      );
-      expect(defaultMocks.fetcher.fetch).toBeCalledWith(
-        IssuerConfigFetcherFetchConfigResponse.tokenEndpoint,
-        {
-          method: "POST",
-          headers: {
-            DPoP: DpopHeaderCreatorResponse,
-            "content-type": "application/x-www-form-urlencoded",
-            Authorization: "Basic Yzpl"
-          },
-          body:
-            "client_id=c&grant_type=authorization_code&code_verifier=a&code=someCode&redirect_uri=d"
-        }
-      );
-      expect(
-        defaultMocks.tokenSaver.saveTokenAndGetSession
-      ).toHaveBeenCalledWith("userId", "idToken", "accessToken", undefined);
+      ).rejects.toThrowError("There was a problem creating a session.");
     });
   });
 });
